fix: ajustei os cards e deixei lista dinamica para os jogos
